p2p: Optimize registry, move stream stuff around
License: MIT
